Fixed BoulderSolver and WeirdosSolver Cleanup
