--- odinmain/src/main/kotlin/me/odinmain/utils/skyblock/dungeon/tiles/RoomData.kt
@@ -1,29 +1,0 @@
-package me.odinmain.utils.skyblock.dungeon.tiles
-
-import com.google.gson.*
-import com.google.gson.reflect.TypeToken
-import java.lang.reflect.Type
-
-data class RoomData(
-    val name: String,
-    val type: RoomType,
-    val cores: List<Int>,
-    val crypts: Int,
-    val secrets: Int,
-    val trappedChests: Int,
-)
-
-class RoomDataDeserializer : JsonDeserializer<RoomData> {
-    override fun deserialize(json: JsonElement?, typeOfT: Type?, context: JsonDeserializationContext?): RoomData {
-        val jsonObject = json?.asJsonObject
-        val name = jsonObject?.get("name")?.asString ?: ""
-        val type = context?.deserialize(jsonObject?.get("type"), RoomType::class.java) ?: RoomType.NORMAL
-        val coresType = object : TypeToken<List<Int>>() {}.type
-        val cores = context?.deserialize<List<Int>>(jsonObject?.get("cores"), coresType).orEmpty()
-        val crypts = jsonObject?.get("crypts")?.asInt ?: 0
-        val secrets = jsonObject?.get("secrets")?.asInt ?: 0
-        val trappedChests = jsonObject?.get("trappedChests")?.asInt ?: 0
-
-        return RoomData(name, type, cores, crypts, secrets, trappedChests)
-    }
-}
